pyproject.toml

[tool.hatch.version]
path = "fedflow/__init__.py"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "fedflow"
description = "Federated FeatureCloud.ai workflows on remote machines"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv==1.2.1",
  "fabric==3.2.2",
  "FeatureCloud==0.1.0",
  "httpx==0.28.1",
  "pydantic==2.12.5",
  "tomli_w==1.2.0",
]


[project.scripts]
fedflow = "fedflow.cli:main"
fcauto = "fedflow.fcauto:main"


[tool.hatch.build.targets.sdist]
only-include = [
  "fedflow/",
]


[tool.hatch.build.targets.wheel]
packages = ["fedflow"]


[tool.hatch.envs.dev]
dependencies = [
  "pytest",
  "pytest-cov",
]


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = [
    "integration: tests running VMs",
]

